Refactor the code to MVC pattern for the game
Refactor the Wall class to be the GameModel class. It holds all the
current game information.

Refactor the GameBoard class and split it into GameView.fxml and the
GameController class

Implemented the ResourceHandler class to handle all the resource
request by the game. Removes the absolute path dependency for
resources.